requirements.txt
pygame
mutagen
pyserial
serial
websocket-client
gpiozero